Update main.js
Significantly cut down on repeated code, by just sticking them into a function and calling the function. Additionally, I'm clearing images in fullscreen now, and I have it so that the speech recognition doesn't stop automatically in fullscreen. We might need a command again to have it stop in this case, but un-full screening and refreshing I believe works.